refactor : Profile 관련 Component 공통화
